# swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mock-Airbnb-REST-API",
    "description": "RESTful API built with node.js, express.js and MongoDB that exposes Airbnb data of Athens in the 2015-2016 time period. It uses JWT tokens encapsulated in cookies for authentication and provides endpoints for different roles (admins, hosts, simple users).",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    }
  },
  "servers": [
    {
      "url": "https://airbnbapi.sidirope.info"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwtCookie"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/admin/allusers/{format}": {
      "get": {
        "tags": ["admin"],
        "parameters": [
          {
            "in": "path",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": ["json", "xml"]
            },
            "required": true,
            "description": "format of the results",
            "example": "json"
          }
        ],
        "description": "Retrieve all users in Json or XML format",
        "responses": {
          "200": {
            "description": "successfull operation"
          },
          "400": {
            "description": "invalid format specidied"
          },
          "403": {
            "description": "No permission to access this resource"
          },
          "500": {
            "description": "server error"
          }
        },
        "servers": [
          {
            "url": "https://airbnbapi.sidirope.info"
          }
        ]
      },
      "servers": [
        {
          "url": "https://airbnbapi.sidirope.info"
        }
      ]
    },
...
    "/admin/validatehost/{hostID}": {
      "patch": {
        "tags": ["admin"],
        "parameters": [
          {
            "in": "path",
            "name": "hostID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the host to be validated",
            "example": "5e9dcc9e0d970d27fd97da52"
          }
        ],
        "description": "Endpoint for admins to validate unvalidated hosts",
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid format supplied"
          },
          "403": {
            "description": "No permission to access this resource"
          },
          "500": {
            "description": "Server error"
          }
        },
        "servers": [
          {
            "url": "https://airbnbapi.sidirope.info"
          }
        ]
      },
      "servers": [
        {
          "url": "https://airbnbapi.sidirope.info"
        }
      ]
    },
...
    "/admin/listings/{format}": {
      "get": {
        "tags": ["admin"],
        "parameters": [
          {
            "in": "path",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": ["json", "xml"]
            },
            "required": true,
            "description": "format of the results",
            "example": "json"
          }
        ],
        "description": "Return all listings in Json or XML format.",
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid format supplied"
          },
          "403": {
            "description": "No permission to access this resource"
          },
          "500": {
            "description": "Server error"
          }
        },
        "servers": [
          {
            "url": "https://airbnbapi.sidirope.info"
          }
        ]
      },
      "servers": [
        {
          "url": "https://airbnbapi.sidirope.info"
        }
      ]
    },
...
    "/admin/pendinghosts": {
      "get": {
        "tags": ["admin"],
        "description": "Retrieve all currently pending hosts",
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid format supplied"
          },
          "403": {
            "description": "No permission to access this resource"
          },
          "404": {
            "description": "no pending host exists"
          },
          "500": {
            "description": "Server error"
          }
        },
        "servers": [
          {
            "url": "https://airbnbapi.sidirope.info"
          }
        ]
      },
      "servers": [
        {
          "url": "https://airbnbapi.sidirope.info"
        }
      ]
    },
...
  }
}